Fix dir encryption and clean up code
Move main file check into node as we do not want to rely on : in file name.
Use encrypt file system call for directories.
Call removeExtraStreams for ads in files and folders from restore metadata before restoring meta data so that readonly is false before we start removing the streams.